Create MetricalDuration module
