+++ Package.swift
@@ -9,7 +9,8 @@ let package = Package(
         .library(name: "Articulations", targets: ["Articulations"]),
         .library(name: "Dynamics", targets: ["Dynamics"]),
         .library(name: "Pitch", targets: ["Pitch"]),
-        .library(name: "Rhythm", targets: ["Rhythm"]),
+        .library(name: "Rhythm", targets: ["MetricalDuration"]),
+        .library(name: "MetricalDuration", targets: ["MetricalDuration"]),
         //.library(name: "MetricalDuration", targets: ["MetricalDuration"]),
         .library(name: "Tempo", targets: ["Tempo"]),
         .library(name: "Meter", targets: ["Meter"]),
@@ -25,7 +26,8 @@ let package = Package(
         .target(name: "Articulations", dependencies: []),
         .target(name: "Dynamics", dependencies: ["Destructure", "Structure"]),
         .target(name: "Pitch", dependencies: ["Math", "StructureWrapping"]),
-        .target(name: "Rhythm", dependencies: ["Math", "DataStructures", "Structure"]),
+        .target(name: "Rhythm", dependencies: ["MetricalDuration"]),
+        .target(name: "MetricalDuration", dependencies: ["Math", "DataStructures", "Structure"]),
         .target(name: "Tempo", dependencies: ["Rhythm"]),
         .target(name: "Meter", dependencies: ["Rhythm", "Tempo"]),
         .target(name: "MusicModel", dependencies: [
@@ -44,6 +46,7 @@ let package = Package(
         .testTarget(name: "DynamicsTests", dependencies: ["Dynamics"]),
         .testTarget(name: "PitchTests", dependencies: ["Pitch"]),
         .testTarget(name: "RhythmTests", dependencies: ["Rhythm"]),
+        .testTarget(name: "MetricalDurationTests", dependencies: ["MetricalDuration"]),
         .testTarget(name: "TempoTests", dependencies: ["Tempo"]),
         .testTarget(name: "MeterTests", dependencies: ["Meter"]),
         .testTarget(name: "MusicModelTests", dependencies: ["MusicModel"])
