+++ Sources/MetricalDuration/Beats.swift
