+++ Sources/MusicModel/Model.Builder.swift
@@ -11,6 +11,7 @@ import Algebra
 import DataStructures
 import Math
 import Pitch
+import MetricalDuration
 import Rhythm
 import Tempo
 import Meter
@@ -249,14 +250,6 @@ extension Model {
         return Builder()
     }
 }
-
-//extension Model.Builder: CustomStringConvertible {
-//    
-//    public var description: String {
-//        return "\(values)"
-//    }
-//}
-
 
 /// - TODO: Move to `dn-m/Rhythm`.
 extension Rhythm {
