+++ Sources/Tempo/Tempo.swift
@@ -6,8 +6,8 @@
 //
 //
 
-// FIXME: Update to `MetricalDuration`
 import Rhythm
+import MetricalDuration
 
 /// Model of a `Tempo`.
 public struct Tempo {
